+++ lib/nunes/subscribers/active_job.rb
@@ -13,16 +13,15 @@ module Nunes
 
       def perform(start, ending, transaction_id, payload)
         runtime = ((ending - start) * 1_000).round
-        job = payload[:job].class.to_s.underscore
+        job = payload[:job].class.to_s
 
         timing "active_job.#{job}.perform", runtime
       end
 
       def enqueue(start, ending, transaction_id, payload)
-        job = payload[:job].class.to_s.underscore
+        job = payload[:job].class.to_s
         increment "active_job.#{job}.enqueue"
       end
     end
   end
 end
-