+++ test/rails_app/app/controllers/posts_controller.rb
@@ -1,12 +1,6 @@
 class PostsController < ApplicationController
-  # Use fake post for controller as I don't want active record to mingle here.
-  Post = Struct.new(:title)
-
   def index
-    @posts = [
-      Post.new('First'),
-      Post.new('Second'),
-    ]
+    @posts = Post.all
   end
 
   def some_data