+++ test/rails_app/app/jobs/spam/detector_job.rb
@@ -1,0 +1,11 @@
+module Spam
+  class DetectorJob < ActiveJob::Base
+    queue_as :default
+
+    def perform(*posts)
+      posts.detect do |post|
+        post.title.include?("Buy watches cheap!")
+      end
+    end
+  end
+end